// cypress/support/types/daniTestTypes.ts
export interface ResponseBodyCards {
  id: string;
  badges: {};
    location: boolean;
    votes: number;
    viewingMemberVoted: boolean;
    subscribed: boolean;
    fogbugz: string;
    checkItems: number;
    checkItemsChecked: number;
    checkItemsEarliestDue: null;
    comments: number;
    attachments: number;
    description: boolean;
    due: null;
    dueComplete: boolean;
    start: null;
  };
  checkItemStates: [];
  closed: boolean;
  dueComplete: boolean;
  dateLastActivity: string;
  desc: string;
  descData: {
    emoji: {};
  };
  due: null;
  dueReminder: null;
  email: null;
  idBoard: string;
  idChecklists: [];
  idList: string;
  idMembers: any[];
  idMembersVoted: any[];
  idShort: number;
  idAttachmentCover: null;
  labels: any[];
  idLabels: any[];
  manualCoverAttachment: boolean;
  name: string;
  pos: number;
  shortLink: string;
  shortUrl: string;
  start: null;
  subscribed: boolean;
  url: string;
  cover: {
    idAttachment: null;
    color: null;
    idUploadedBackground: null;
    size: string;
    brightness: string;
    idPlugin: null;
  };
  isTemplate: boolean;
  cardRole: null;
  attachments: any[];
  stickers: any[];
  limits: {};
}